+++ packages/react/src/useUnmount.ts
@@ -1,0 +1,6 @@
+import ComponentSetup from './ComponentSetup';
+
+export default function useUnmount(fn: () => void) {
+  const setup = ComponentSetup.getCurrent<ComponentSetup>();
+  setup.unmountFns.push(fn);
+}